% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.network.R
\name{contrast.inits}
\alias{contrast.inits}
\title{contrast.network.deviance.plot <- function(result){
  deviance <- result$deviance
  dev_vector <- c(t(deviance$dev_arm))
  dev_vector <- dev_vector[!is.na(dev_vector)]
  plot(seq(sum(result$network$na)), dev_vector, xlab = "Arm", ylab = "Residual Deviance", main = "Per-arm residual deviance")
}}
\usage{
contrast.inits(network, n.chains)
}
\description{
#' Make a leverage plot
#'
#' This function makes a leverage vs. square root of residual deviance plot
#' 
#' @param result Object created by \code{\link{network.run}} function
#' @export
}
\details{
network.leverage.plot <- function(result){
  deviance <- result$deviance
  dev <- sqrt(apply(deviance$dev_arm, 1, mean, na.rm = TRUE))
  leverage <- apply(deviance$leverage, 1, mean, na.rm = TRUE)
  plot(dev, leverage, xlim = c(0, max(c(dev, 2.5))), ylim = c(0, max(c(leverage,4))),
       xlab = "Square root of residual deviance", ylab = "Leverage", main = "Leverage versus residual deviance")
  mtext("Per-study mean per-datapoint contribution")
}
}
