% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.autocorr.diag}
\alias{network.autocorr.diag}
\title{Use coda package to find autocorrelation diagnostics}
\usage{
network.autocorr.diag(result, lags = c(0, 1, 5, 10, 50), extra.pars = NULL,
  only.pars = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{lags}{a vector of lags at which to calculate the autocorrelation}

\item{extra.pars}{extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{parameters that user wants to plot only}
}
\description{
Use coda package to find autocorrelation diagnostics
}
\examples{
Outcomes <- blocker[["Outcomes"]]
Study <- blocker[["Study"]]
Treat <- blocker[["Treat"]]
N <- blocker[["N"]]
network <- network.data(Outcomes, Study, Treat, N = N, response = "binomial")
result <- network.run(network)
network.autocorr.diag(result, only.pars = "d")
}
