% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.autocorr.plot}
\alias{network.autocorr.plot}
\title{Use coda package to plot autocorrelation plot}
\usage{
network.autocorr.plot(result, extra.pars = NULL, only.pars = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{extra.pars}{extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{parameters that user wants to plot only}
}
\description{
Use coda package to plot autocorrelation plot
}
\examples{
#cardiovascular
Study <- cardiovascular[["Study"]]
Treat <- cardiovascular[["Treat"]]
Outcomes <- cardiovascular[["Outcomes"]]
N <- cardiovascular[["N"]]
network <- network.data(Outcomes, Study, Treat, N, response = "multinomial")
result <- network.run(network)
network.autocorr.plot(result, only.pars = "d")
}
