% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.covariate.plot}
\alias{network.covariate.plot}
\title{Make a covariate plot}
\usage{
network.covariate.plot(result, base.treatment = NULL,
  comparison.treatment = NULL, base.category = NULL,
  comparison.category = NULL, covariate.name = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{base.treatment}{base treatment for relative effect}

\item{comparison.treatment}{treatment comparing against base treatment}

\item{base.category}{base category for multinomial data}

\item{comparison.category}{comparison category for multinomial data}

\item{covariate.name}{A vector of covariate names naming of the covariate that goes into x axis label}
}
\description{
Make a covariate plot of how the relative effect changes as the covariate value changes. Plot is created for each one of the covariate.
User needs to specify one base treatment and one comparison treatment to make this plot (base category and comparison category is needed for multinomial).
It then uses the \code{\link{relative.effects}} to calculate the correct relative effect.
2.5\% and 97.5\% C.I. are drawn along with the median value.
}
\examples{
########### certolizumab (with covariate)
Outcomes <- certolizumab[["Outcomes"]]
N <- certolizumab[["N"]]
Study <- certolizumab[["Study"]]
Treat <- certolizumab[["Treat"]]
covariate <- certolizumab[["covariate"]]
Treat.order <- certolizumab[["Treat.order"]]
network <- network.data(Outcomes, Study, Treat, N=N, response="binomial", Treat.order,
covariate = covariate, hy.prior = list("dhnorm", 0, 9.77))
result <- network.run(network)
network.covariate.plot(result, base.treatment = "Placebo", comparison.treatment = "CZP",
covariate.name = "Disease Duration")
}
