% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.deviance.plot}
\alias{network.deviance.plot}
\title{make a deviance plot}
\usage{
network.deviance.plot(result)
}
\arguments{
\item{result}{object created by \code{network.run} function}
}
\description{
This makes a deviance plot which plots residual deviance (dev_arm) vs. all the arms for each study.
}
\examples{
Outcomes <- blocker[["Outcomes"]]
Study <- blocker[["Study"]]
Treat <- blocker[["Treat"]]
N <- blocker[["N"]]
network <- network.data(Outcomes, Study, Treat, N = N, response = "binomial")
result <- network.run(network)
network.deviance.plot(result)
}
