% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.forest.plot}
\alias{network.forest.plot}
\title{Draws forest plot}
\usage{
network.forest.plot(result, level = 0.95, xlim = NULL, ylim = NULL,
  refline = 0)
}
\description{
Draws forest plot using odds ratio of pooled treatment effect
}
