% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.gelman.diag}
\alias{network.gelman.diag}
\title{Use coda package to find gelman-diagnostic diagnostics}
\usage{
network.gelman.diag(result, extra.pars = NULL, only.pars = NULL)
}
\arguments{
\item{result}{object created by \code{\link{network.run}} function}

\item{extra.pars}{extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{parameters that user wants to plot only}
}
\description{
Use coda package to find gelman-diagnostic diagnostics
}
\examples{
network <- with(statins, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial",
 Treat.order = c("Placebo", "Statin"), covariate = covariate, covariate.type = "discrete")
})
result <- network.run(network)
network.gelman.diag(result, extra.pars = "Eta")
}
