% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.leverage.plot}
\alias{network.leverage.plot}
\title{make a leverage plot}
\usage{
network.leverage.plot(result)
}
\arguments{
\item{result}{object created by \code{\link{network.run}} function}
}
\description{
Make a leverage vs. square root of residual deviance plot
}
