% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.rank.tx.plot}
\alias{network.rank.tx.plot}
\title{Creates a treatment rank plot}
\usage{
network.rank.tx.plot(result, txnames = NULL, catnames = NULL,
  legend.position = c(1, 1))
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{txnames}{treatment names used in creating legend}

\item{catnames}{category names. Only used in multinomial.}

\item{legend.position}{x,y position of the legend}
}
\description{
Creates a treatment rank plot
}
\examples{
network <-with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
network.rank.tx.plot(result, txnames = c("a", "b"))
}
\seealso{
\code{\link{rank.tx}}
}
