% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesplit.network.R
\name{nodesplit.network.data}
\alias{nodesplit.network.data}
\title{Assessing consistency using node-splitting model}
\usage{
nodesplit.network.data(Outcomes, Study, Treat, N = NULL, SE = NULL,
  response = NULL, type = "random")
}
\arguments{
\item{Outcomes}{Arm-level outcomes. If it is a multinomial response, the matrix would be arms (row) by multinomial categories (column). If it is binomial or normal, it would be a vector.}

\item{Study}{A vector of study indicator for each arm}

\item{Treat}{A vector of treatment indicator for each arm. Treatments should have positive integer values starting from 1 to total number of treatments. In a study, lowest number is taken as the baseline treatment.}

\item{N}{A vector of total number of observations in each arm. Used for binomial and multinomial responses.}

\item{SE}{A vector of standard error for each arm. Used only for normal response.}

\item{response}{Specification of the outcomes type. Must specify one of the following: "normal", "binomial", or "multinomial".}

\item{type}{Type of model fitted: either "random" for random effects model or "fixed" for fixed effects model. Default is "random".}
}
\description{
This function is used to make node-splitting inconsistency model. The structure is similar to the function \code{\link{network.data}}
}
\references{
S. Dias, N.J. Welton, D.M. Caldwellb, A.E. Ades (2010), \emph{Checking consistency in mixed treatment}, Statistics in Medicine 29(7-8, Sp. Iss. SI): 932-944. [\url{https://doi.org/10.1002/sim.3767}]
}
