% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{plot.network.result}
\alias{plot.network.result}
\title{plot traceplot and posterior density of the result}
\usage{
\method{plot}{network.result}(x, ...)
}
\arguments{
\item{x}{result object created by \code{\link{network.run}} function}

\item{...}{additional arguments affecting the plot produced}
}
\description{
Use plotting function in coda package to plot mcmc.list object
}
\examples{
network <- with(statins, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial",
 Treat.order = c("Placebo", "Statin"), covariate = covariate, covariate.type = "discrete")
})
result <- network.run(network)
plot(result, only.pars = "sd")
}
