% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{rank.tx}
\alias{rank.tx}
\title{Creates a treatment rank table}
\usage{
rank.tx(result)
}
\arguments{
\item{result}{object created by \code{network.run} function}
}
\value{
This makes a table of ranking for each treament. Each number in the cell represents a probability certain treatment was in such rank.
This table is also stored as an output from \code{network.run}.
}
\description{
Creates a treatment rank table
}
\examples{
network <- with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
rank.tx(result)
}
\seealso{
\code{\link{network.rank.tx.plot}}
}
