% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{relative.effects}
\alias{relative.effects}
\title{Find relative effects for different base treatment and comparison treatments}
\usage{
relative.effects(result, base.treatment = NULL,
  comparison.treatments = NULL, base.category = NULL,
  comparison.categories = NULL, covariate = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{base.treatment}{base treatment user wants for the relative effects. Base treatment is initially set by \code{Treat.order} parameter in \code{network.data} (first one in the list). If set to null, default is to use base treatment.}

\item{comparison.treatments}{treatments that user wants to compare against base treatment. If set to null, all the treatments besides base treatment is considered as comparison treatments.}

\item{base.category}{base category user wants for the relative effects. This is only used for multinomial data.}

\item{comparison.categories}{category that user wants to compare against base.category}

\item{covariate}{covariate value at which to compute relative effects.}
}
\value{
This returns a mcmc.list sample of relative effects for the base treatment specified. This allows user to obtain relative effects of different base.treatment after the sampling has been done.
For a simple summary use relative.effects.table.
}
\description{
Find relative effects for different base treatment and comparison treatments
}
\examples{
#We can fit two different models with different base treatment and we can
#obtain same relative effects estimate using this function
#parkinsons
Outcomes <- parkinsons[["Outcomes"]]
Study <- parkinsons[["Study"]]
Treat <- parkinsons[["Treat"]]
SE <- parkinsons[["SE"]]
network <- network.data(Outcomes, Study, Treat, SE = SE, response = "normal")
result <- network.run(network)

Treat.order <- c(2,1,3,4,5)
network2 <- network.data(Outcomes, Study, Treat, SE = SE, response = "normal",
Treat.order = Treat.order)
result2 <- network.run(network2)

summary(result2)
summary(relative.effects(result2, base.treatment = 2))

#This also works for comparing different base.category for multinomial.
#We fit two different models and compare the estimates again.
#cardiovascular

Study <- cardiovascular[["Study"]]
Treat <- cardiovascular[["Treat"]]
Outcomes <- cardiovascular[["Outcomes"]]
N <- cardiovascular[["N"]]
network3 <- network.data(Outcomes, Study, Treat, N, response = "multinomial")
result3 <- network.run(network3)

Outcomes <- Outcomes[,c(2,1,3)]
network4 <- network.data(Outcomes, Study, Treat, N, response = "multinomial")
result4 <- network.run(network4)

summary(result3)
summary(relative.effects(result4, base.category = 2))
}
\seealso{
\code{\link{relative.effects.table}}
}
