% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{relative.effects.table}
\alias{relative.effects.table}
\title{Make a summary table for relative effects}
\usage{
relative.effects.table(result, summary_stat = "mean", base.category = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{summary_stat}{specifies what type of statistics user wants. Options are: "mean", "quantile", "sd", "p-value".
Quantile use c(0.025, 0.5, 0.975). P-value is the probability relative effect is less than 0.}

\item{base.category}{specifies for which base category user wants for the summary. Used only for multinoimal.}
}
\description{
Make a summary table for relative effects
}
\examples{
#cardiovascular
network <- with(cardiovascular,{
 network.data(Outcomes, Study, Treat, N, response = "multinomial")
})
result <- network.run(network)
exp(relative.effects.table(result)) #look at odds ratio instead of log odds ratio
}
\seealso{
\code{\link{relative.effects}}
}
