% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{variance.tx.effects}
\alias{variance.tx.effects}
\title{Calculates correlation matrix for multinomial heterogeneity parameter.}
\usage{
variance.tx.effects(result)
}
\arguments{
\item{result}{object created by \code{network.run} function}
}
\description{
Calculates correlation matrix from the variance matrix for heterogeneity parameter. Only used for multinomial.
}
\examples{
#cardiovascular
Study <- cardiovascular[["Study"]]
Treat <- cardiovascular[["Treat"]]
Outcomes <- cardiovascular[["Outcomes"]]
N <- cardiovascular[["N"]]
network <- network.data(Outcomes, Study, Treat, N, response = "multinomial")
result <- network.run(network)
variance.tx.effects(result)
}
