% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.autocorr.diag}
\alias{network.autocorr.diag}
\title{Generate autocorrelation diagnostics using coda package}
\usage{
network.autocorr.diag(result, lags = c(0, 1, 5, 10, 50), extra.pars = NULL,
  only.pars = NULL)
}
\arguments{
\item{result}{object created by \code{\link{network.run}} function}

\item{lags}{a vector of lags at which to calculate the autocorrelation}

\item{extra.pars}{extra parameters that the user wants to display other than the default parameters.}

\item{only.pars}{parameters that user wants to display. This gets rids of other default parameters user doesn't want to show.}
}
\description{
Using the coda package, generate autocorrelation diagnostics. User can specify lags and parameters to display. 
Note that to display extra parameters, user needs to first specify parameters in extra.pars.save parameter in \code{\link{network.run}} function so the model saves those parameters of interest.
}
\examples{
network <- with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
network.autocorr.diag(result, only.pars = "d")
}
