% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.gelman.plot}
\alias{network.gelman.plot}
\title{Use coda package to plot gelman-diagnostic plot}
\usage{
network.gelman.plot(result, extra.pars = NULL, only.pars = NULL)
}
\arguments{
\item{result}{object created by \code{network.run} function}

\item{extra.pars}{extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{parameters that user wants to plot only}
}
\description{
Use coda package to plot gelman-diagnostic plot
}
\examples{
#blocker
network <- with(blocker,{
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
network.gelman.plot(result, only.pars = "d")
}
