% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{summary.network.result}
\alias{summary.network.result}
\title{summarize result run by \code{network.run}}
\usage{
\method{summary}{network.result}(object, ...)
}
\arguments{
\item{object}{result object created by \code{network.run} function}

\item{...}{additional arguments affecting the summary produced
#' @examples
network <- with(statins, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial",
 Treat.order = c("Placebo", "Statin"), covariate = covariate, covariate.type = "discrete")
})
result <- network.run(network)
summary(result)}
}
\description{
Use summary function in \code{coda} to summarize mcmc.list object
}
